+++ algebra/integer_magnitude_test.cc
@@ -868,8 +868,8 @@ TEST_CASE("num_bits") {
 }
 
 TEST_CASE("popcount") {
-    for (uint i: {0, 4, 31231, -3123121})
-        REQUIRE(integer(i).popcount() == std::popcount(i));
+    for (uint i: {0u, 4u, 31231u, static_cast<uint>(-3123121)})
+        REQUIRE(integer(i).popcount() == static_cast<size_t>(std::popcount(i)));
 }
 
 TEST_CASE("<<") {
